Restructure extension: create `LegendSpecFits` extension
