--- ext/LegendMakieMeasurementsExt.jl
@@ -1,430 +1,0 @@
-# This file is a part of LegendMakie.jl, licensed under the MIT License (MIT).
-
-module LegendMakieMeasurementsExt
-
-    import LegendMakie
-
-    import LaTeXStrings
-    import Makie
-    import Measurements
-    import StatsBase
-    import Unitful
-
-    import LegendMakie: pt, aoecorrectionplot!, energycalibrationplot!
-
-    # Default color palette
-    function get_default_color(i::Int)
-        colors = Makie.wong_colors()
-        colors[(i - 1) % end + 1]
-    end
-
-    function round_wo_units(x::Unitful.RealOrRealQuantity; digits::Int=2)
-        Unitful.unit(x) == Unitful.NoUnits ? round(x; digits) : round(Unitful.unit(x), x; digits)
-    end
-
-    # extends LegendMakie.default_xlims in LegendMakieExt.jl
-    # for the cases where μ and σ have uncertainties
-    function LegendMakie.default_xlims(report::NamedTuple{(:f_fit, :h, :μ, :σ, :gof),<:Tuple{Any,Any,M,M,Any}}
-        ) where {M <: Union{Measurements.Measurement, Unitful.Quantity{<:Measurements.Measurement}}}
-        Unitful.ustrip.(Measurements.value.((report.μ - 5*report.σ, report.μ + 5*report.σ)))
-    end
-
-    # function to compose labels when errorbars are scaled
-    function label_errscaling(xerrscaling::Real, yerrscaling::Real = 1)
-        result = String[]
-        xerrscaling != 1 && push!(result, "x-Error ×$(xerrscaling)")
-        yerrscaling != 1 && push!(result, "y-Error ×$(yerrscaling)")
-        isempty(result) ? "" : " ($(join(result,", ")))"
-    end
-
-
-    Makie.@recipe(AoECorrectionPlot, report) do scene
-        Makie.Attributes(
-            color = (LegendMakie.AchatBlue,0.5),
-            markercolor = :black
-        )
-    end
-    
-    # Needed for creatings legend using Makie recipes
-    # https://discourse.julialang.org/t/makie-defining-legend-output-for-a-makie-recipe/121567
-    function Makie.get_plots(p::AoECorrectionPlot)
-        return p.plots
-    end
-    
-    function Makie.plot!(p::AoECorrectionPlot{<:Tuple{NamedTuple{(:par, :f_fit, :x, :y, :gof, :e_unit, :label_y, :label_fit)}}})
-        report = p.report[]
-        Makie.lines!(p, 0:1:3000, report.f_fit ∘ Measurements.value, color = p.color, label = LaTeXStrings.latexstring("\\fontfamily{Roboto}" * report.label_fit))
-        Makie.errorbars!(p, report.x, Measurements.value.(report.y), Measurements.uncertainty.(report.y), color = p.markercolor)
-        Makie.scatter!(p, report.x, Measurements.value.(report.y), color = p.markercolor, label = "Compton band fits: Gaussian $(report.label_y)(A/E)")
-        p
-    end
-
-
-    Makie.@recipe(EnergyCalibrationPlot, report, additional_pts) do scene
-        Makie.Attributes(
-            color = LegendMakie.AchatBlue,
-            plot_ribbon = true,
-            xerrscaling = 1,
-            yerrscaling = 1
-        )
-    end
-
-    # Needed for creatings legend using Makie recipes
-    # https://discourse.julialang.org/t/makie-defining-legend-output-for-a-makie-recipe/121567
-    function Makie.get_plots(p::EnergyCalibrationPlot)
-        return p.plots
-    end
-    
-    function Makie.plot!(p::EnergyCalibrationPlot{<:Tuple{<:NamedTuple{(:par, :f_fit, :x, :y, :gof)}, <:NamedTuple}})
-        
-        report = p.report[]
-        additional_pts = p.additional_pts[]
-        xerrscaling = p.xerrscaling[]
-        yerrscaling = p.yerrscaling[]
-
-        # plot fit
-        xfit = 0:1:1.2*Measurements.value(maximum(report.x))
-        yfit = report.f_fit.(xfit)
-        yfit_m = Measurements.value.(yfit)
-        Makie.lines!(xfit, yfit_m, label = "Best Fit" * (!isempty(report.gof) ? " (p = $(round(report.gof.pvalue, digits=2))| χ²/ndf = $(round(report.gof.chi2min, digits=2)) / $(report.gof.dof))" : ""), color = p.color)
-        if p.plot_ribbon[]
-            Δyfit = Measurements.uncertainty.(yfit)
-            Makie.band!(xfit, yfit_m .- Δyfit, yfit_m .+ Δyfit, color = (p.color[], 0.2))
-        end
-        
-        # scatter points with error bars
-        xvalues = Measurements.value.(report.x)
-        yvalues = Measurements.value.(report.y)
-        Makie.errorbars!(p, xvalues, yvalues, Measurements.uncertainty.(report.x) .* xerrscaling, direction = :x, color = :black)
-        Makie.errorbars!(p, xvalues, yvalues, Measurements.uncertainty.(report.y) .* yerrscaling, color = :black)
-        Makie.scatter!(p, xvalues, yvalues, marker = :circle, color = :black, label = "Data" * label_errscaling(xerrscaling, yerrscaling))
-        
-        # plot additional points
-        if !isempty(additional_pts)
-            xvalues = Measurements.value.(additional_pts.x)
-            yvalues = Measurements.value.(additional_pts.y)
-            Makie.errorbars!(p, xvalues, yvalues, Measurements.uncertainty.(additional_pts.x) .* xerrscaling, direction = :x, color = :black)
-            Makie.errorbars!(p, xvalues, yvalues, Measurements.uncertainty.(additional_pts.y) .* yerrscaling, color = :black)
-            Makie.scatter!(p, xvalues, yvalues, marker = :circle, color = :silver, strokewidth = 1, strokecolor = :black, label = "Data not used for fit" * label_errscaling(xerrscaling, yerrscaling))
-        end
-
-        p
-    end
-
-
-    function LegendMakie.lplot!(
-            report::NamedTuple{(:par, :f_fit, :x, :y, :gof)};
-            additional_pts::NamedTuple = NamedTuple(),
-            xlims = (0, 1.2*Measurements.value(maximum(report.x))), ylims = nothing,
-            xlabel = "Energy (ADC)", ylabel = "Energy (calibrated)", title::AbstractString = "",
-            show_residuals::Bool = true, plot_ribbon::Bool = true, legend_position = :lt,
-            xerrscaling::Real = 1, yerrscaling::Real = 1, row::Int = 1, col::Int = 1,
-            watermark::Bool = true, final::Bool = !isempty(title), kwargs...
-        )
-        
-        fig = Makie.current_figure()
-            
-        g = Makie.GridLayout(fig[row,col])
-        ax = Makie.Axis(g[1,1],
-            title = title,
-            limits = (xlims, ylims),
-            xlabel = xlabel, ylabel = ylabel,
-        )
-        
-        LegendMakie.energycalibrationplot!(ax, report, additional_pts; plot_ribbon, xerrscaling, yerrscaling)
-        legend_position != :none && Makie.axislegend(ax, position = legend_position)
-        
-        if !isempty(report.gof) && show_residuals
-
-            ax.xticklabelsize = 0
-            ax.xticksize = 0
-            ax.xlabel = ""
-
-            ax2 = Makie.Axis(g[2,1], yticks = -3:3:3, limits = (xlims,(-5,5)), xlabel = xlabel, ylabel = "Residuals (σ)")
-            LegendMakie.residualplot!(ax2, (x = Measurements.value.(report.x), residuals_norm = report.gof.residuals_norm))
-            # add the additional points
-            if !isempty(additional_pts)
-                Makie.scatter!(ax2, Measurements.value.(additional_pts.x), additional_pts.residuals_norm, 
-                        marker = :circle, color = :silver, strokewidth = 1, strokecolor = :black)
-            end
-
-            # link axis and put plots together
-            Makie.linkxaxes!(ax, ax2)
-            Makie.rowgap!(g, 0)
-            Makie.rowsize!(g, 1, Makie.Auto(4))
-
-            # align ylabels
-            yspace = maximum(Makie.tight_yticklabel_spacing!, (ax, ax2))
-            ax.yticklabelspace = yspace
-            ax2.yticklabelspace = yspace
-        end
-
-        # add watermarks
-        Makie.current_axis!(ax)
-        watermark && LegendMakie.add_watermarks!(; final, kwargs...)
-
-        fig
-    end
-
-
-    function LegendMakie.lplot!(
-            report::NamedTuple{(:par, :f_fit, :x, :y, :gof, :e_unit, :type)}; 
-            additional_pts::NamedTuple = NamedTuple(), kwargs...
-        )
-
-        if report.type != :cal
-            @warn "Unknown calibration type $(report.type), no plot generated."
-            return Makie.current_figure()
-        end
-                 
-        LegendMakie.lplot!(
-            report[(:par, :f_fit, :x, :y, :gof)],
-            additional_pts = if !isempty(additional_pts)
-                # strip the units from the additional points
-                μ_strip = Unitful.unit(first(additional_pts.μ)) != Unitful.NoUnits ? Unitful.ustrip.(report.e_unit, additional_pts.μ) : additional_pts.μ
-                p_strip = Unitful.unit(first(additional_pts.peaks)) != Unitful.NoUnits ? Unitful.ustrip.(report.e_unit, additional_pts.peaks) : additional_pts.peaks    
-                μ_cal = report.f_fit.(μ_strip)
-                (x = μ_strip, y = p_strip, residuals_norm = (Measurements.value.(μ_cal) .- Measurements.value.(p_strip))./ Measurements.uncertainty.(μ_cal))
-            else
-                NamedTuple()
-            end,
-            xlabel = "Energy (ADC)", ylabel = "Energy ($(report.e_unit))", xlims = (0, 1.1*Measurements.value(maximum(report.x))); kwargs...
-        )
-    end
-
-    function LegendMakie.lplot!(
-            report::NamedTuple{(:par, :f_fit, :x, :y, :gof, :e_unit, :qbb, :type)};
-            additional_pts::NamedTuple = NamedTuple(), xlims = (0,3000), title::AbstractString = "",
-            kwargs...
-        )
-        
-        if report.type != :fwhm
-            @warn "Unknown calibration type $(report.type), no plot generated."
-            return Makie.current_figure()
-        end
-
-        fig = LegendMakie.lplot!(
-            report[(:par, :f_fit, :x, :y, :gof)],
-            additional_pts = if !isempty(additional_pts)
-                fwhm_cal = report.f_fit.(Unitful.ustrip.(additional_pts.peaks))
-                (x = Unitful.ustrip.(report.e_unit, additional_pts.peaks), y = Unitful.ustrip.(report.e_unit, additional_pts.fwhm),
-                    residuals_norm = (Measurements.value.(fwhm_cal .- Unitful.ustrip.(report.e_unit, additional_pts.fwhm))) ./ Measurements.uncertainty.(fwhm_cal))
-            else
-                NamedTuple()
-            end,
-            xlabel = "Energy ($(report.e_unit))", ylabel = "FWHM ($(report.e_unit))",
-            legend_position = :none; xlims, title, kwargs...
-        )
-        
-        ax = first(fig.content)
-        Makie.current_axis!(ax)
-        qbb = Unitful.ustrip(report.e_unit, Measurements.value(report.qbb))
-        Δqbb = Unitful.ustrip(report.e_unit, Measurements.uncertainty(report.qbb))
-        Makie.hlines!(ax, [qbb], color = LegendMakie.CoaxGreen, label = LaTeXStrings.latexstring("\\fontfamily{Roboto}Q_{\\beta \\beta}:" * " $(round(Unitful.ustrip(report.e_unit, report.qbb), digits=2))\\;\\text{$(report.e_unit)}"))
-        Makie.band!(ax, range(xlims..., length = 2), fill(qbb - Δqbb, 2), fill(qbb + Δqbb, 2), color = (LegendMakie.CoaxGreen, 0.2))
-        Makie.axislegend(ax, position = :lt)
-        
-        fig
-    end
-
-    LegendMakie.lplot!(report::NamedTuple{(:peak, :n_before, :n_after, :sf, :before, :after)}; kwargs...) = LegendMakie.lplot!(report.after; kwargs...)
-        
-    function LegendMakie.lplot!(
-            report::NamedTuple{((:survived, :cut, :sf, :bsf))};
-            xlims::Tuple{<:Real, <:Real} = extrema(first(report.survived.h.edges)), 
-            xticks = (ceil(first(xlims)/10)*10):10:(floor(last(xlims)/10)*10),
-            ylims = nothing, row::Int = 1, col::Int = 1, xlabel = "Energy (keV)",
-            title::AbstractString = "", yscale = Makie.log10, show_residuals::Bool = true, kwargs...
-        )
-            
-        fig = Makie.current_figure()
-        g = Makie.GridLayout(fig[row,col])
-        
-        if isnothing(ylims)
-            ylim_max = 1.5 * max(
-                Measurements.value(report.survived.f_fit(report.survived.v.μ)), 
-                Measurements.value(report.cut.f_fit(report.cut.v.μ)),
-                maximum(report.survived.h.weights), maximum(report.cut.h.weights)
-            )
-            ylim_max = ifelse(iszero(ylim_max), nothing, max(1.2e2, ylim_max))
-            ylim_min = 0.5 * min(minimum(filter(x -> x > 0, report.survived.h.weights)), minimum(filter(x -> x > 0, report.cut.h.weights)))
-            ylims = (ylim_min, ylim_max)
-        end
-
-        ax = Makie.Axis(g[1,1], yticks = (exp10.(0:10), "1" .* join.(fill.("0", 0:10))), 
-            ylabel = "Counts / $(round(step(first(report.survived.h.edges)), digits=2)) keV", 
-            limits = (xlims, ylims); yscale, xlabel, title, xticks)
-        
-        Makie.plot!(ax, report.survived.h, color = (:gray, 0.5), label = "Data Survived", fillto = 0.5)
-        Makie.lines!(ax, range(xlims..., length = 1000), x -> report.survived.f_fit(x) * step(first(report.survived.h.edges)), color = :black, label = "Best Fit" * (!isempty(report.survived.gof) ? " (p = $(round(report.survived.gof.pvalue, digits=2)))" : ""))
-        Makie.plot!(ax, report.cut.h, color = (:lightgray, 0.5), label = "Data Cut", fillto = 0.5)
-        Makie.lines!(ax, range(xlims..., length = 1000), x -> report.cut.f_fit(x) * step(first(report.cut.h.edges)), color = (:gray, 0.5), label = "Best Fit" * (!isempty(report.cut.gof) ? " (p = $(round(report.cut.gof.pvalue, digits=2)))" : ""))
-        Makie.axislegend(ax, position = :lt)
-
-        if !isempty(report.survived.gof) && show_residuals
-            ax2 = Makie.Axis(g[2,1], limits = (extrema(first(report.survived.h.edges)), (-5,5)), xlabel = xlabel, xticks = xticks, ylabel = "Residuals (σ)")
-            LegendMakie.residualplot!(ax2,(x = report.survived.gof.bin_centers, residuals_norm = report.survived.gof.residuals_norm), color = (:black, 0.5))
-            Makie.scatter!(report.cut.gof.bin_centers, report.cut.gof.residuals_norm, color = (:darkgray, 0.7))
-
-            ax.xticklabelsize = 0
-            ax.xticksize = 0
-            ax.xlabel = ""        
-
-            Makie.linkxaxes!(ax, ax2)
-            Makie.rowgap!(g, 0)
-            Makie.rowsize!(g, 1, Makie.Auto(3))
-
-            # align ylabels
-            yspace = maximum(Makie.tight_yticklabel_spacing!, (ax, ax2))
-            ax.yticklabelspace = yspace
-            ax2.yticklabelspace = yspace
-
-            if row == 1
-                Makie.current_axis!(ax)
-                LegendMakie.add_watermarks!(final = true)
-            end
-        end
-        
-        fig
-    end
-
-    # SiPM Gaussian Mixture fit
-    function LegendMakie.lplot!(
-            report::NamedTuple{(:h_cal, :f_fit, :f_fit_components, :min_pe, :max_pe, :bin_width, :n_mixtures, :n_pos_mixtures, :peaks, :positions, :μ, :gof)};
-            show_peaks::Bool = true, show_residuals::Bool = true, show_components::Bool = true, show_label::Bool = true,
-            xlims = extrema(first(report.h_cal.edges)), title::AbstractString = "", yscale = Makie.log10,
-            ylims = yscale == Makie.log10 ? (10, maximum(report.h_cal.weights)*4) : (0, maximum(report.h_cal.weights)*1.2),
-            xlabel = "Peak amplitudes (P.E.)", ylabel = "Counts", xerrscaling = 1,
-            row::Int = 1, col::Int = 1, xticks = Makie.automatic, yticks = Makie.automatic,
-            legend_position = :rt, watermark::Bool = true, final::Bool = !isempty(title), kwargs...
-        )
-
-        fig = Makie.current_figure()
-
-        # create plot
-        g = Makie.GridLayout(fig[row,col])
-        ax = Makie.Axis(g[1,1], 
-            title = title, titlefont = :bold, titlesize = 16pt, 
-            limits = (xlims, ylims); xlabel, ylabel, xticks, yticks, yscale
-        )
-
-        Makie.hist!(ax, StatsBase.midpoints(first(report.h_cal.edges)), weights = report.h_cal.weights, bins = first(report.h_cal.edges), color = LegendMakie.DiamondGrey, label = "Amplitudes", fillto = 1e-2)
-        Makie.lines!(range(xlims..., length = 1000), x -> report.f_fit(x), linewidth = 2, color = :black, label = ifelse(show_label, "Best Fit" * (!isempty(report.gof) ? " (p = $(round(report.gof.pvalue, digits=2)))" : ""), nothing))
-
-        # show individual components of the Gaussian mixtures
-        if show_components
-            for i in eachindex(report.μ)
-                f = Base.Fix2(report.f_fit_components, i)
-                Makie.lines!(
-                    range(extrema(first(report.h_cal.edges))..., length = 1000), 
-                    x -> f(x), 
-                    #color = report.f_components.colors[component], 
-                    label = ifelse(show_label && i == firstindex(report.μ) , "Mixture Components", nothing),
-                    color = (get_default_color(i), 0.5),
-                    linestyle = :dash,
-                    linewidth = 2
-                )
-            end
-        end
-
-        # show peak positions as vlines
-        if show_peaks
-            for (i, p) in enumerate(report.positions)
-                Makie.vlines!([Measurements.value(p)], label = "$(report.peaks[i]) P.E. [$(report.n_pos_mixtures[i]) Mix.]" * label_errscaling(xerrscaling,1), linewidth = 1.5)
-                Makie.band!(ax, [(Measurements.value(p) .+ (-1,1) .* xerrscaling .* Measurements.uncertainty(p))...], ylims..., alpha = 0.5)
-            end
-        end
-
-        # add residuals
-        if !isempty(report.gof) && show_residuals
-
-            ax.xticklabelsize = 0
-            ax.xticksize = 0
-            ax.xlabel = ""
-
-            ax2 = Makie.Axis(g[2,1],
-                xlabel = "Peak amplitudes (P.E.)", ylabel = "Residuals (σ)",
-                xticks = xticks, yticks = -3:3:3, limits = (xlims,(-5,5))
-            )
-            LegendMakie.residualplot!(ax2, (x = report.gof.bin_centers, residuals_norm = report.gof.residuals_norm))
-
-            # link axis and put plots together
-            Makie.linkxaxes!(ax, ax2)
-            Makie.rowgap!(g, 0)
-            Makie.rowsize!(g, 1, Makie.Auto(3))
-
-            # align ylabels
-            yspace = maximum(Makie.tight_yticklabel_spacing!, (ax, ax2))
-            ax.yticklabelspace = yspace
-            ax2.yticklabelspace = yspace
-        end
-
-        legend_position != :none && Makie.axislegend(ax, backgroundcolor = (:white, 0.7), framevisible = true, framecolor = :lightgray, position = legend_position)
-
-        # add watermarks
-        Makie.current_axis!(ax)
-        watermark && LegendMakie.add_watermarks!(; final, kwargs...)
-
-        fig
-    end
-
-    # filter optimzation plots
-    function LegendMakie.lplot!(
-            report::NamedTuple{(:x, :minx, :y, :miny)};
-            title::AbstractString = "", xunit = Unitful.unit(first(report.x)), yunit = Unitful.unit(first(report.y)),
-            xlabel = "", ylabel = "", xlegendlabel = xlabel, ylegendlabel = ylabel, 
-            xlims = Unitful.ustrip.(xunit, extrema(report.x) .+ (-1, 1) .* (report.x[2] - report.x[1])),
-            legend_position = :lt, watermark::Bool = true, final::Bool = !isempty(title), kwargs...
-        )
-
-        fig = Makie.current_figure()
-
-        ax = Makie.Axis(fig[1,1];
-            title, limits = (xlims, nothing),
-            xlabel = (xlabel * ifelse(xunit == Unitful.NoUnits, "", " ($xunit)")) |> typeof(xlabel),
-            ylabel = (ylabel * ifelse(yunit == Unitful.NoUnits, "", " ($yunit)")) |> typeof(ylabel)
-        )
-
-        Makie.errorbars!(ax, Unitful.ustrip.(xunit, report.x), Unitful.ustrip.(yunit, Measurements.value.(report.y)), Unitful.ustrip.(yunit, Measurements.uncertainty.(report.y)))
-        Makie.scatter!(ax, Unitful.ustrip.(xunit, report.x), Unitful.ustrip.(yunit, Measurements.value.(report.y)), label = ylegendlabel)
-        Makie.hlines!(ax, [Unitful.ustrip(yunit, Measurements.value(report.miny))], color = :red, label = "Min. $(ylegendlabel) $(round_wo_units(report.miny, digits = 2)) $yunit ($(xlegendlabel): $(report.minx))")
-        Makie.band!(ax, [xlims...], Unitful.ustrip.(yunit, (Measurements.value(report.miny) .+ (-1,1) .* Measurements.uncertainty(report.miny)))..., color = (:red,0.2))
-        
-        legend_position != :none && Makie.axislegend(ax, position = legend_position)
-
-        Makie.current_axis!(ax)
-        watermark && LegendMakie.add_watermarks!(; final, kwargs...)
-        fig
-    end
-
-    function LegendMakie.lplot!(report::NamedTuple{(:rt, :min_enc, :enc_grid_rt, :enc)}; kwargs...)
-        LegendMakie.lplot!(
-            (x = report.enc_grid_rt, minx = report.rt, y = report.enc, miny = report.min_enc);
-            xlabel = "Rise Time", ylabel = "ENC (ADC)", xlegendlabel = "RT", ylegendlabel = "ENC noise", kwargs...
-        )
-    end
-
-    function LegendMakie.lplot!(report::NamedTuple{(:ft, :min_fwhm, :e_grid_ft, :fwhm)}; kwargs...)
-        LegendMakie.lplot!(
-            (x = report.e_grid_ft, minx = report.ft, y = report.fwhm, miny = report.min_fwhm);
-            xlabel = "Flat-Top Time", ylabel = "FWHM", xlegendlabel = "FT", ylegendlabel = "FWHM", kwargs...
-        )
-    end
-
-    function LegendMakie.lplot!(report::NamedTuple{(:wl, :min_sf, :a_grid_wl_sg, :sfs)}; kwargs...)
-        LegendMakie.lplot!(
-            (x = report.a_grid_wl_sg, minx = report.wl, y = report.sfs, miny = report.min_sf);
-            xlabel = "Window length", ylabel = "SEP survival fraction", xlegendlabel = "WL", ylegendlabel = "SF", kwargs...
-        )
-    end
-
-    function LegendMakie.lplot!(report::NamedTuple{(:wl, :min_obj, :gain, :res_1pe, :pos_1pe, :threshold, :a_grid_wl_sg, :obj, :report_simple, :report_fit)}; kwargs...)
-        LegendMakie.lplot!(
-            (x = report.a_grid_wl_sg, minx = report.wl, y = report.obj, miny = report.min_obj);
-            xlabel = "Window length",
-            ylabel = LaTeXStrings.latexstring("\\fontfamily{Roboto} Objective\\; \$\\frac{\\sqrt{\\sigma \\cdot \\text{threshold}}}{\\text{gain}}\$"), 
-            xlegendlabel = "WL", ylegendlabel = "Obj", kwargs...
-        )
-    end
-
-end
