+++ src/lplot.jl
@@ -47,6 +47,3 @@ function add_legend_logo! end
 function add_juleana_logo! end
 function add_text! end
 function add_watermarks! end
-
-# auxiliary functions
-function default_xlims end
